Remove 1.19.3 support.
